Implement new authorisation for users

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -32,12 +32,12 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            $this->isAble('store', User::class);
+        if ($this->isAble('store', User::class)) {
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $e) {
-            return $this->error('This action is unauthorized.', 403);
-        }
+        } 
+            
+        return $this->error('This action is unauthorized.', 403);
+
     }
 
     /**
@@ -60,16 +60,16 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('update', $user);
+            if($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
-
+            return $this->error('This action is unauthorized.', 403);
+        
         } catch (ModelNotFoundException $e) {
             return $this->error('User not found', 404);
-        } catch (AuthorizationException $e) {
-            return $this->error('This action is unauthorized.', 403);
         }
     }
 
@@ -78,11 +78,13 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('replace', $user);
+            if($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
+            return $this->error('This action is unauthorized.', 403);
 
         } catch (ModelNotFoundException $e) {
             return $this->error('User not found', 404);
@@ -96,10 +98,15 @@ class UserController extends ApiController
     {
         try {
             $user = User::findOrFail($user_id);
-            $this->isAble('delete', $user);
-            $user->delete();
+            
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
 
-            return $this->ok('User deleted successfully.');
+                return $this->ok('User deleted successfully.');
+            }
+
+            return $this->error('This action is unauthorized.', 403);
+            
         } catch (ModelNotFoundException $e) {
             return $this->error('User not found', 404);
         }
